hw12 makefile changed
