+++ .golangci.yml
@@ -11,8 +11,10 @@ linters-settings:
 
 issues:
   exclude-dirs:
-    - vendor
-    - /home/runner/go
+    # от ошибок линтеров на исходники golang в пайплайне github - Error: ../../../../../../opt/hostedtoolcache/go/1.23.4
+    - /opt/hostedtoolcache/go/
+    # от ошибок линтеров на исходники используемых библиотек в пайплайне github - Error: ../../../../go/pkg/mod/github.com/jackc/pgx@v3.6.2+incompatible/conn.go:1491:6: softErr declared and not used (typecheck)
+    - ../../../../go/pkg/mod/
   exclude-rules:
     - path: _test\.go
       linters:
